refactor: separated company list in dedicated component
