+++ src/shared/constants/store/homePage/compayLogos.ts
@@ -1,0 +1,42 @@
+export const COMPANIES_LOGOS = [
+  {
+    width: 104,
+    bgPositionX: 0,
+    url: "/list/photography/scanners/epson",
+  },
+  {
+    width: 50,
+    bgPositionX: -105,
+    url: "/",
+  },
+  {
+    width: 50,
+    bgPositionX: -156,
+    url: "/",
+  },
+  {
+    width: 44,
+    bgPositionX: -207,
+    url: "/list/smartphones/apple-iphone",
+  },
+  {
+    width: 47,
+    bgPositionX: -252,
+    url: "/list/smartphones/xiaomi",
+  },
+  {
+    width: 54,
+    bgPositionX: -300,
+    url: "/",
+  },
+  {
+    width: 55,
+    bgPositionX: -355,
+    url: "/",
+  },
+  {
+    width: 98,
+    bgPositionX: -411,
+    url: "/list/pc-laptops/laptops/asus",
+  },
+];
